asserts in _release

+++ src/rex/dev/common.h
@@ -1,7 +1,7 @@
 __device__ float fst(float2 s) { return s.x; }
 __device__ float scn(float2 s) { return s.y; }
 
-__device__ void i2shift(int fid, /**/ float *d) {
+__device__ void i2shift(int fid, /**/ int *d) {
     /* fragment id to coordinate shift */
     enum {X, Y, Z};
     d[X] = - ((fid +     2) % 3 - 1) * XS;
